+++ compiler-core/src/language_server/code_action.rs
@@ -10,7 +10,7 @@ use crate::{
         AssignName, AssignmentKind, CallArg, ImplicitCallArgOrigin, Pattern, SrcSpan, TypedExpr,
         TypedPattern, TypedRecordUpdateArg,
     },
-    build::Module,
+    build::{Located, Module},
     line_numbers::LineNumbers,
     parse::extra::ModuleExtra,
     type_::{self, error::ModuleSuggestion, FieldMap, ModuleValueConstructor, Type, TypedCallArg},
@@ -933,5 +933,161 @@ fn suggest_imports(
             location,
             suggestions,
         })
+    }
+}
+
+pub fn code_action_add_missing_patterns(
+    module: &Module,
+    params: &CodeActionParams,
+    error: &Option<Error>,
+    actions: &mut Vec<CodeAction>,
+) {
+    let uri = &params.text_document.uri;
+    let Some(Error::Type { errors, .. }) = error else {
+        return;
+    };
+    let missing_patterns = errors
+        .iter()
+        .filter_map(|error| match error {
+            type_::Error::InexhaustiveCaseExpression { location, missing } => {
+                Some((*location, missing))
+            }
+            _ => None,
+        })
+        .collect_vec();
+
+    if missing_patterns.is_empty() {
+        return;
+    }
+
+    let line_numbers = LineNumbers::new(&module.code);
+
+    for (location, missing) in missing_patterns {
+        let range = src_span_to_lsp_range(location, &line_numbers);
+        if !overlaps(params.range, range) {
+            return;
+        }
+
+        let mut edits = Vec::new();
+
+        let Some(Located::Expression(TypedExpr::Case {
+            clauses, subjects, ..
+        })) = module.find_node(location.start)
+        else {
+            continue;
+        };
+
+        // Find the start of the line. We can't just use the start of the case
+        // expression for cases like:
+        //
+        //```gleam
+        // let value = case a {}
+        //```
+        //
+        // Here, the start of the expression is part-way through the line, meaning
+        // we think we are more indented than we actually are
+        //
+        let mut indent_size = 0;
+        let line_start = *line_numbers
+            .line_starts
+            .get(range.start.line as usize)
+            .expect("Line number should be valid");
+        let chars = module.code.chars();
+        let mut chars = chars.skip(line_start as usize);
+        // Count indentation
+        while chars.next() == Some(' ') {
+            indent_size += 1;
+        }
+
+        let indent = " ".repeat(indent_size);
+
+        // Insert the missing patterns just after the final clause, or just before
+        // the closing brace if there are no clauses
+        let insert_span = clauses
+            .last()
+            .map(|clause| SrcSpan {
+                start: clause.location.end,
+                end: clause.location.end,
+            })
+            .unwrap_or(SrcSpan {
+                start: location.end - 1,
+                end: location.end - 1,
+            });
+
+        let insert_range = src_span_to_lsp_range(insert_span, &line_numbers);
+
+        for pattern in missing {
+            let new_text = format!("\n{indent}  {pattern} -> todo");
+            edits.push(TextEdit {
+                range: insert_range,
+                new_text,
+            })
+        }
+
+        // Add a newline + indent after the last pattern if there are no clauses
+        //
+        // This improves the generated code for this case:
+        // ```gleam
+        // case True {}
+        // ```
+        // This produces:
+        // ```gleam
+        // case True {
+        //   True -> todo
+        //   False -> todo
+        // }
+        // ```
+        // Instead of:
+        // ```gleam
+        // case True {
+        //   True -> todo
+        //   False -> todo}
+        // ```
+        //
+        if clauses.is_empty() {
+            let last_subject_location = subjects
+                .last()
+                .expect("Case expressions have at least one subject")
+                .location()
+                .end;
+
+            // Find the opening brace of the case expression
+
+            // Calculate the number of characters from the start of the line to the end of the
+            // last subject, to skip, so we can find the opening brace.
+            // That is: the location we want to get to, minus the start of the line which we skipped to begin with,
+            // minus the number we skipped for the indent, minus one more because we go one past the end of indentation
+            let num_to_skip = last_subject_location - line_start - indent_size as u32 - 1;
+            let chars = chars.skip(num_to_skip as usize);
+            let mut start_brace_location = last_subject_location;
+            for char in chars {
+                start_brace_location += 1;
+                if char == '{' {
+                    break;
+                }
+            }
+
+            let range = src_span_to_lsp_range(
+                SrcSpan::new(start_brace_location, insert_span.start),
+                &line_numbers,
+            );
+
+            // Remove any blank spaces/lines between the start brace and end brace
+            edits.push(TextEdit {
+                range,
+                new_text: String::new(),
+            });
+
+            edits.push(TextEdit {
+                range: insert_range,
+                new_text: format!("\n{indent}"),
+            })
+        }
+
+        CodeActionBuilder::new("Add missing patterns")
+            .kind(CodeActionKind::QUICKFIX)
+            .changes(uri.clone(), edits)
+            .preferred(true)
+            .push_to(actions);
     }
 }