+++ compiler-core/src/parse.rs
@@ -1626,7 +1626,7 @@ where
         Ok(Pattern::Constructor {
             location: SrcSpan { start, end },
             arguments: args,
-            module: module.map(|(_, n, _)| n),
+            module: module.map(|(start, n, end)| (n, SrcSpan { start, end })),
             name,
             spread,
             constructor: Inferred::Unknown,
@@ -3535,7 +3535,7 @@ functions are declared separately from types.";
         let target = match name.as_str() {
             "erlang" => Target::Erlang,
             "javascript" => Target::JavaScript,
-            _ => return parse_error(ParseErrorType::UnknownAttribute, SrcSpan::new(start, end)),
+            _ => return parse_error(ParseErrorType::UnknownTarget, SrcSpan::new(start, end)),
         };
 
         let _ = self.expect_one(&Token::Comma)?;