+++ compiler-core/src/type_/error.rs
@@ -984,6 +984,7 @@ pub enum UnknownValueConstructorError {
 pub fn convert_get_value_constructor_error(
     e: UnknownValueConstructorError,
     location: SrcSpan,
+    module_location: Option<SrcSpan>,
 ) -> Error {
     match e {
         UnknownValueConstructorError::Variable {
@@ -998,7 +999,7 @@ pub fn convert_get_value_constructor_error(
         },
 
         UnknownValueConstructorError::Module { name, suggestions } => Error::UnknownModule {
-            location,
+            location: module_location.unwrap_or(location),
             name,
             suggestions,
         },