+++ compiler-core/src/type_/expression.rs
@@ -2268,8 +2268,9 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             TypedExpr::ModuleSelect {
                 module_alias,
                 label,
+                location,
                 ..
-            } => (Some(module_alias), label),
+            } => (Some((module_alias, location)), label),
 
             TypedExpr::Var { name, .. } => (None, name),
 
@@ -2282,8 +2283,14 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
 
         let value_constructor = self
             .environment
-            .get_value_constructor(module.as_ref(), &name)
-            .map_err(|e| convert_get_value_constructor_error(e, location))?
+            .get_value_constructor(module.as_ref().map(|(module, _)| module), &name)
+            .map_err(|e| {
+                convert_get_value_constructor_error(
+                    e,
+                    location,
+                    module.as_ref().map(|(_, location)| *location),
+                )
+            })?
             .clone();
 
         // It must be a record with a field map for us to be able to update it
@@ -2636,7 +2643,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 // Potentially this could be improved later
                 match self
                     .get_field_map(&fun)
-                    .map_err(|e| convert_get_value_constructor_error(e, location))?
+                    .map_err(|e| convert_get_value_constructor_error(e, location, None))?
                 {
                     // The fun has a field map so labelled arguments may be present and need to be reordered.
                     Some(field_map) => field_map.reorder(&mut args, location)?,
@@ -2954,7 +2961,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         // Check to see if the function accepts labelled arguments
         let field_map = self
             .get_field_map(&fun)
-            .map_err(|e| convert_get_value_constructor_error(e, location))
+            .map_err(|e| convert_get_value_constructor_error(e, location, None))
             .and_then(|field_map| {
                 match field_map {
                     // The fun has a field map so labelled arguments may be present and need to be reordered.
@@ -3335,7 +3342,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         let mut compiler = Compiler::new(self.environment, Arena::new());
         let mut arena = PatternArena::new();
 
-        let subject_variable = compiler.new_variable(subject.clone());
+        let subject_variable = compiler.subject_variable(subject.clone());
 
         let mut rows = Vec::with_capacity(1);
 
@@ -3392,7 +3399,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
 
         let subject_variables = subject_types
             .iter()
-            .map(|t| compiler.new_variable(t.clone()))
+            .map(|t| compiler.subject_variable(t.clone()))
             .collect_vec();
 
         let mut rows = Vec::with_capacity(clauses.iter().map(Clause::pattern_count).sum::<usize>());