+++ compiler-core/src/type_/pattern.rs
@@ -459,8 +459,14 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
 
                 let cons = self
                     .environment
-                    .get_value_constructor(module.as_ref(), &name)
-                    .map_err(|e| convert_get_value_constructor_error(e, location))?;
+                    .get_value_constructor(module.as_ref().map(|(module, _)| module), &name)
+                    .map_err(|e| {
+                        convert_get_value_constructor_error(
+                            e,
+                            location,
+                            module.as_ref().map(|(_, location)| *location),
+                        )
+                    })?;
 
                 match cons.field_map() {
                     // The fun has a field map so labelled arguments may be present and need to be reordered.