+++ compiler-core/src/type_/tests/exhaustiveness.rs
@@ -1,4 +1,6 @@
-use crate::{assert_module_error, assert_no_warnings, assert_warning, assert_with_module_error};
+use crate::{
+    assert_error, assert_module_error, assert_no_warnings, assert_warning, assert_with_module_error,
+};
 
 #[test]
 fn whatever() {
@@ -1121,6 +1123,149 @@ pub fn main() {
   case wibble {
     mod.Wobble -> Nil
   }
+}
+"
+    );
+}
+
+// The following few tests all verify that the compiler provides useful errors
+// when there are no case arms, instead of just suggesting `_` as it did previously.
+#[test]
+fn empty_case_of_bool() {
+    assert_error!(
+        "
+let b = True
+case b {}
+"
+    );
+}
+
+#[test]
+fn empty_case_of_custom_type() {
+    assert_module_error!(
+        "
+type Wibble { Wibble Wobble Wubble }
+pub fn main() {
+  let wibble = Wobble
+  case wibble {}
+}
+"
+    );
+}
+
+#[test]
+fn empty_case_of_list() {
+    assert_error!(
+        "
+let list = []
+case list {}
+"
+    );
+}
+
+#[test]
+fn empty_case_of_int() {
+    assert_error!(
+        "
+let num = 24
+case num {}
+"
+    );
+}
+
+#[test]
+fn empty_case_of_float() {
+    assert_error!(
+        "
+let age = 10.6
+case age {}
+"
+    );
+}
+
+#[test]
+fn empty_case_of_string() {
+    assert_error!(
+        r#"
+let name = "John Doe"
+case name {}
+"#
+    );
+}
+
+#[test]
+fn empty_case_of_multi_pattern() {
+    assert_error!(
+        "
+let a = Ok(1)
+let b = True
+case a, b {}
+"
+    );
+}
+
+#[test]
+fn inexhaustive_multi_pattern() {
+    assert_error!(
+        "
+let a = Ok(1)
+let b = True
+case a, b {
+  Error(_), _ -> Nil
+}
+"
+    );
+}
+
+#[test]
+fn inexhaustive_multi_pattern2() {
+    assert_error!(
+        "
+let a = Ok(1)
+let b = True
+case a, b {
+  Ok(1), True -> Nil
+}
+"
+    );
+}
+
+#[test]
+fn inexhaustive_multi_pattern3() {
+    assert_error!(
+        "
+let a = Ok(1)
+let b = True
+case a, b {
+  _, False -> Nil
+}
+"
+    );
+}
+
+#[test]
+fn inexhaustive_multi_pattern4() {
+    assert_error!(
+        "
+let a = 12
+let b = 3.14
+let c = False
+case a, b, c {
+  1, 2.0, True -> Nil
+}
+"
+    );
+}
+
+#[test]
+fn inexhaustive_multi_pattern5() {
+    assert_error!(
+        "
+let a = 12
+let b = 3.14
+let c = False
+case a, b, c {
+  12, _, False -> Nil
 }
 "
     );